avoid int-unsigned int warnings
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@6330 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/operator/2order/dgprimaloperator.hh
@@ -655,7 +655,7 @@ protected:
     void resizeCaches(const int numDofs) const
     {
       // resize caches 
-      if(tau_.size() != numDofs) 
+      if(tau_.size() != (unsigned int)(numDofs)) 
       { 
         tau_.resize(numDofs);
         tauNeigh_.resize(numDofs);
